chat-control-wrapper
I made control wrapper that appears when you choose a chat and click the icon, where you can see some options to control this chat (del, edit, export). I made the basic functionalities and it is just basic, so it is not the final code, just to test the logic.